+++ test/dummy/app/models/js_configuration.rb
@@ -1,0 +1,8 @@
+class JsConfiguration < ActiveRecord::Base
+  include DiscoApp::Concerns::RendersAssets
+
+  belongs_to :shop, class_name: 'DiscoApp::Shop'
+
+  renders_assets :js_assets, assets: 'test.js', triggered_by: 'locale'
+
+end