+++ test/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160521135510) do
+ActiveRecord::Schema.define(version: 20160530160739) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -108,11 +108,24 @@ ActiveRecord::Schema.define(version: 20160521135510) do
   add_index "disco_app_subscriptions", ["plan_id"], name: "index_disco_app_subscriptions_on_plan_id", using: :btree
   add_index "disco_app_subscriptions", ["shop_id"], name: "index_disco_app_subscriptions_on_shop_id", using: :btree
 
+  create_table "js_configurations", force: :cascade do |t|
+    t.integer "shop_id", limit: 8
+    t.string  "label",             default: "Default"
+    t.string  "locale",            default: "en"
+  end
+
   create_table "products", force: :cascade do |t|
     t.integer  "shop_id",    limit: 8
     t.jsonb    "data"
     t.datetime "created_at",           null: false
     t.datetime "updated_at",           null: false
+  end
+
+  create_table "widget_configurations", force: :cascade do |t|
+    t.integer "shop_id",          limit: 8
+    t.string  "label",                      default: "Default"
+    t.string  "locale",                     default: "en"
+    t.string  "background_color",           default: "#FFFFFF"
   end
 
   add_foreign_key "disco_app_application_charges", "disco_app_shops", column: "shop_id"
@@ -122,5 +135,7 @@ ActiveRecord::Schema.define(version: 20160521135510) do
   add_foreign_key "disco_app_recurring_application_charges", "disco_app_subscriptions", column: "subscription_id"
   add_foreign_key "disco_app_sessions", "disco_app_shops", column: "shop_id", on_delete: :cascade
   add_foreign_key "disco_app_subscriptions", "disco_app_plan_codes", column: "plan_code_id"
+  add_foreign_key "js_configurations", "disco_app_shops", column: "shop_id"
   add_foreign_key "products", "disco_app_shops", column: "shop_id"
+  add_foreign_key "widget_configurations", "disco_app_shops", column: "shop_id"
 end